Added doc comments to these database and entity classes

+++ phase1/src/usecases/art/Appraiser.java
@@ -13,7 +13,7 @@ public class Appraiser {
      * @param art => the art piece to appraise
      * @return float representing the market price of given art piece
      */
-    public float appraiseArt(ArtManager library, Art art){
+    public double appraiseArt(ArtManager library, Art art){
         return art.getPrice();
     }
 }
